move the extract metadata concern to submission process step file
